Create fist vew

+++ app/src/main/java/com/example/shopinglist/domain/GetShopItemUseCase.kt
@@ -1,7 +1,7 @@
 package com.example.shopinglist.domain
 
 class GetShopItemUseCase(private val repository: ShopListRepository) {
-    fun getShopItem(shopItemId: Int): ShopItem {
+    fun getShopItem(shopItemId: Long): ShopItem {
       return repository.getShopItem(shopItemId)
     }
 }
