+++ app/src/main/java/com/example/shopinglist/domain/ShopListRepository.kt
@@ -6,6 +6,6 @@ interface ShopListRepository {
     fun deleteItem(shopItem: ShopItem)
     fun addShopItem(shopItem: ShopItem)
     fun editShopItem(shopItem: ShopItem)
-    fun getShopItem(shopItemId: Int): ShopItem
+    fun getShopItem(shopItemId: Long): ShopItem
     fun getList(): LiveData<List<ShopItem>>
 }
